Fix new warning in t/op/mkdir.t.
Removing a directory that doesn't exist causes a warning, and if
the warning appears before the test preamble, that causes the test
to fail.  So check whether the directory exists before trying to
remove it.

+++ t/op/mkdir.t
@@ -10,7 +10,7 @@ plan tests => 22;
 
 unless (eval {
     require File::Path;
-    File::Path::rmtree('blurfl');
+    File::Path::rmtree('blurfl') if -d 'blurfl';
     1
 }) {
     diag("$0 may fail if its temporary directory remains from a previous run");
